rename *Schema to  *Graphic components
