+++ src/components/graphics/Arrow.vue
@@ -32,10 +32,10 @@
 </template>
 
 <script>
-import SchemaComponent from './SchemaComponent.vue'
+import GraphicComponent from './GraphicComponent.vue'
 
 export default {
-  extends: SchemaComponent
+  extends: GraphicComponent
 }
 </script>
