+++ src/components/graphics/GraphicComponent.vue
