+++ src/components/graphics/MessageCircle.vue
@@ -85,11 +85,11 @@
 </template>
 
 <script>
-import SchemaComponent from './SchemaComponent.vue'
+import GraphicComponent from './GraphicComponent.vue'
 import MessagePopover from './MessagePopover.vue'
 
 export default {
-  extends: SchemaComponent,
+  extends: GraphicComponent,
 
   components: {
     MessagePopover
