+++ src/components/transaction-graphic/TransferTransactionSchema copy.vue
@@ -84,7 +84,7 @@
 
 <script>
 import helper from '../../helper'
-import SchemaComponent from '../graphics/SchemaComponent.vue'
+import GraphicComponent from '../graphics/GraphicComponent.vue'
 import AccountIcon from '../graphics/AccountIcon.vue'
 import MessageCircle from '../graphics/MessageCircle.vue'
 import MosaicsCircle from '../graphics/MosaicsCircle.vue'
@@ -92,7 +92,7 @@ import NativeMosaicCircle from '../graphics/NativeMosaicCircle.vue'
 import Arrow from '../graphics/Arrow.vue'
 
 export default {
-  extends: SchemaComponent,
+  extends: GraphicComponent,
 
   components: {
     AccountIcon,
