+++ src/config/pages/transaction-detail.json
@@ -12,7 +12,7 @@
         "schema": [
             {
                 "layoutOptions": "full-width",
-                "type": "TransactionSchema",
+                "type": "TransactionGraphicWidget",
                 "managerGetter": "transaction/transactionSchema"
             },
             {
