+++ src/views/Home.vue
@@ -31,7 +31,6 @@
             <b-col xs="12" md="12" lg="6">
                 <RecentTransactionsWidget />
             </b-col>
-            <MosaicListWidget />
         </b-row>
     </b-container>
 </template>
