+++ src/views/PageAssembler.vue
@@ -50,7 +50,7 @@ import BaseInfoWidget from '@/components/widgets/BaseInfoWidget.vue'
 import PriceChartWidget from '@/components/widgets/PriceChartWidget.vue'
 import RecentBlocksWidget from '@/components/widgets/RecentBlocksWidget.vue'
 import RecentTransactionsWidget from '@/components/widgets/RecentTransactionsWidget.vue'
-import TransactionSchema from '@/components/widgets/TransactionSchemaWidget.vue'
+import TransactionGraphicWidget from '@/components/widgets/TransactionGraphicWidget.vue'
 
 export default {
   components: {
@@ -59,7 +59,7 @@ export default {
     PriceChartWidget,
     RecentBlocksWidget,
     RecentTransactionsWidget,
-    TransactionSchema
+    TransactionGraphicWidget
   },
 
   props: {
